Fixed unit tests, made new folder for feature dynamics extraction
