--- tsfresh/feature_extraction/gen_example_timeseries_data.py
@@ -1,286 +1,0 @@
-import pandas as pd
-import numpy as np
-import dask.dataframe as dd
-
-
-def gen_example_timeseries_data(container_type="pandas"):
-
-    assert container_type == "pandas" or container_type == "dask"
-
-    y1 = [
-        "0",
-        "0",
-        "0",
-        "345346",
-        "1356",
-        "135",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "32425436",
-        "0",
-        "0",
-        "345346",
-        "0",
-        "44444444444",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "32425436",
-        "0",
-        "0",
-        "345346",
-        "0",
-        "44444444444",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "32425436",
-        "0",
-        "0",
-        "345346",
-        "0",
-        "44444444444",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "32425436",
-        "0",
-        "0",
-        "345346",
-        "0",
-        "44444444444",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-    ]
-
-    y2 = [
-        "457",
-        "352",
-        "3524",
-        "124532",
-        "24",
-        "24",
-        "214",
-        "21",
-        "46",
-        "42521",
-        "532",
-        "634",
-        "32",
-        "64375",
-        "235",
-        "325",
-        "563323",
-        "6",
-        "32",
-        "532",
-        "52",
-        "57",
-        "324",
-        "643",
-        "32",
-        "436",
-        "34",
-        "57",
-        "34",
-        "65",
-        "643",
-        "34",
-        "346",
-        "43",
-        "54",
-        "8",
-        "4",
-        "43",
-        "537",
-        "543",
-        "43",
-        "56",
-        "32",
-        "34",
-        "32",
-        "5",
-        "65",
-        "43",
-        "435",
-        "54",
-        "7654",
-        "5",
-        "67",
-        "54",
-        "345",
-        "43",
-        "32",
-        "32",
-        "65",
-        "76",
-    ]
-
-    y3 = [
-        "3454",
-        "13452",
-        "23534",
-        "12432",
-        "412432",
-        "324",
-        "43",
-        "5",
-        "64",
-        "356",
-        "3245235",
-        "32",
-        "325",
-        "5467",
-        "657",
-        "235",
-        "234",
-        "34",
-        "2344234",
-        "56",
-        "21435",
-        "214",
-        "1324",
-        "4567",
-        "34232",
-        "132214",
-        "42",
-        "34",
-        "343",
-        "3443",
-        "124",
-        "5477",
-        "36478",
-        "879",
-        "414",
-        "45",
-        "7899",
-        "786",
-        "657",
-        "677",
-        "45645",
-        "3534",
-        "424",
-        "354545",
-        "36645",
-        "67867",
-        "56867",
-        "78876",
-        "5646",
-        "3523",
-        "2434",
-        "324423",
-        "68",
-        "89",
-        "456",
-        "435",
-        "3455",
-        "35443",
-        "24332",
-        "12313",
-    ]
-
-    measurement_id = [
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "1",
-        "2",
-        "2",
-        "2",
-        "2",
-        "2",
-        "2",
-        "3",
-        "3",
-        "3",
-        "3",
-        "3",
-        "3",
-        "4",
-        "4",
-        "4",
-        "4",
-        "4",
-        "4",
-        "5",
-        "5",
-        "5",
-        "5",
-        "5",
-        "5",
-        "6",
-        "6",
-        "6",
-        "6",
-        "6",
-        "6",
-        "7",
-        "7",
-        "7",
-        "7",
-        "7",
-        "7",
-        "8",
-        "8",
-        "8",
-        "8",
-        "8",
-        "8",
-        "9",
-        "9",
-        "9",
-        "9",
-        "9",
-        "9",
-        "10",
-        "10",
-        "10",
-        "10",
-        "10",
-        "10",
-    ]
-
-    ts = pd.DataFrame(
-        {
-            "t": np.repeat([1, 2, 3, 4, 5, 6], 10),
-            "y1": np.asarray(y1, dtype=float),
-            "y2": np.asarray(y2, dtype=float),
-            "y3": np.asarray(y3, dtype=float),
-            "measurement_id": np.asarray(measurement_id, dtype=int),
-        }
-    )
-
-    if container_type == "dask":
-        ts = dd.from_pandas(ts, npartitions=3)
-
-    response = (
-        pd.DataFrame(
-            {
-                "response": np.asarray([0, 1] * 5),
-                "measurement_id": np.asarray(np.arange(1, 11, dtype=int)),
-            }
-        )
-        .set_index("measurement_id")
-        .squeeze()
-    )
-
-    return ts, response
